meta_campaign_builder/requirements.txt
facebook-business==19.0.3
python-dotenv==1.0.1
